Move add_matvec to the misc subpackage
