+++ scikits/cuda/linalg.py
@@ -1656,125 +1656,6 @@ def trace(x_gpu, handle=None):
                        x_gpu.gpudata, incx, one.gpudata, 0)
 
 
-
-add_vec_to_mat_template = Template("""
-#include <pycuda-complex.hpp>
-
-__global__ void addColVecToMat(const ${type} *mat, const ${type} *vec, ${type} *out,
-                               const int32_t n, const int32_t m){
-    const int tx = threadIdx.x;
-    const int ty = threadIdx.y;
-    const int tidx = blockIdx.x * blockDim.x + threadIdx.x;
-    const int tidy = blockIdx.y * blockDim.y + threadIdx.y;
-
-    __shared__ ${type} shared_vec[24];
-
-    if ((ty == 0) & (tidx < n))
-        shared_vec[tx] = vec[tidx];
-    __syncthreads();
-
-    if ((tidy < m) & (tidx < n)) {
-        out[tidx*m+tidy] = mat[tidx*m+tidy] + shared_vec[tx];
-    }
-}
-
-__global__ void addRowVecToMat(const ${type}* mat, const ${type}* vec, ${type}* out,
-                               const int32_t n, const int32_t m){
-    const int tx = threadIdx.x;
-    const int ty = threadIdx.y;
-    const int tidx = blockIdx.x * blockDim.x + threadIdx.x;
-    const int tidy = blockIdx.y * blockDim.y + threadIdx.y;
-
-    __shared__ ${type} shared_vec[24];
-
-    if ((tx == 0) & (tidy < m))
-        shared_vec[ty] = vec[tidy];
-    __syncthreads();
-
-    if ((tidy < m) & (tidx < n)) {
-        out[tidx*m+tidy] = mat[tidx*m+tidy] + shared_vec[ty];
-    }
-}
-""")
-def add_matvec(x_gpu, a_gpu, out=None, stream=None):
-    """
-    Adds a vector to each column/row of the matrix.
-
-    The numpy broadcasting rules apply so this would yield the same result
-    as x_gpu.get() + a_gpu.get() in host-code.
-
-    Parameters
-    ----------
-    x_gpu : pycuda.gpuarray.GPUArray
-        Matrix to add the vector to
-
-    a_gpu : pycuda.gpuarray.GPUArray
-        Vector to add to x_gpu
-
-    out : pycuda.gpuarray.GPUArray (optional)
-        Optional destination matrix
-
-    stream : pycuda.driver.Stream (optional)
-        Optional Stream in which to perform this calculation
-
-    Returns
-    -------
-    out : pycuda.gpuarray.GPUArray
-        result of x_gpu + a_gpu
-    """
-    if a_gpu.shape[0] == x_gpu.shape[0]:
-        axis = 0
-    elif a_gpu.shape[0] == x_gpu.shape[1] or a_gpu.shape[1] == x_gpu.shape[1]:
-        axis = 1
-    else:
-        raise ValueError('Vector length must equal one side of the matrix')
-
-    if x_gpu.dtype == np.float64:
-        datatype = 'double';
-    elif x_gpu.dtype == np.float32:
-        datatype = 'float'
-    if x_gpu.dtype == np.complex64:
-        datatype = 'pycuda::complex<float>';
-    elif x_gpu.dtype == np.complex128:
-        datatype = 'pycuda::complex<double>'
-
-    cache_dir=None
-    tmpl = add_vec_to_mat_template.substitute(type=datatype)
-    mod = SourceModule(tmpl, cache_dir=cache_dir)
-
-    add_row_vec_kernel = mod.get_function('addRowVecToMat')
-    add_col_vec_kernel = mod.get_function('addColVecToMat')
-
-    n, m = np.int32(x_gpu.shape[0]), np.int32(x_gpu.shape[1])
-
-    block = (24, 24, 1)
-    gridx = n // block[0] + 1 * (n % block[0] != 0)
-    gridy = m // block[1] + 1 * (m % block[1] != 0)
-    grid = (gridx, gridy, 1)
-
-    if out is None:
-        alloc = misc._global_cublas_allocator
-        out = gpuarray.empty_like(x_gpu)
-    else:
-        assert out.dtype == x_gpu.dtype
-        assert out.shape == x_gpu.shape
-
-    if x_gpu.flags.c_contiguous:
-        if axis == 0:
-            add_col_vec_kernel(x_gpu, a_gpu, out, n, m,
-                               block=block, grid=grid, stream=stream)
-        elif axis == 1:
-            add_row_vec_kernel(x_gpu, a_gpu, out, n, m,
-                               block=block, grid=grid, stream=stream)
-    else:
-        if axis == 0:
-            add_row_vec_kernel(x_gpu, a_gpu, out, m, n,
-                               block=block, grid=grid, stream=stream)
-        elif axis == 1:
-            add_col_vec_kernel(x_gpu, a_gpu, out, m, n,
-                               block=block, grid=grid, stream=stream)
-    return out
-
 if __name__ == "__main__":
     import doctest
     doctest.testmod()
